packages/design-system/pyproject.toml
[project]
name = "jhg-patterns-ds"
version = "0.1.0"
description = "JHG Design System: theme, components, Flask integration for Databricks apps"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "James Dowzard", email = "james.dowzard@jhg.com.au" }]
dependencies = [
    "pyyaml>=6.0",
    "flask>=2.0",
    "markupsafe>=2.0",
    "pydantic>=2.0",
    "tenacity>=8.0",
]

[project.optional-dependencies]
flask = [
    "flask>=2.0",
    "flask-wtf>=1.0",
    "flask-limiter>=3.0",
    "flask-cors>=4.0",
]
databricks = [
    "databricks-sql-connector>=2.0",
]
graph = [
    "msal>=1.20",
    "requests>=2.28",
]
excel = [
    "openpyxl>=3.0",
]
all = [
    "flask>=2.0",
    "flask-wtf>=1.0",
    "flask-limiter>=3.0",
    "flask-cors>=4.0",
    "databricks-sql-connector>=2.0",
    "msal>=1.20",
    "requests>=2.28",
    "openpyxl>=3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jhg_patterns"]

[tool.hatch.build.targets.wheel.shared-data]
"src/jhg_patterns/ds/templates" = "share/jhg-patterns-ds/templates"
"src/jhg_patterns/ds/assets" = "share/jhg-patterns-ds/assets"
